Initial implementation, ported from the core repository.
Compared to the code from the core repository, the sbt build is in
charge of more operations than was `assemble-cli.sh`. In
particular, it has taken over the tasks of:

* finding the jars to include in `lib/` (most of them are now
  downloaded through ivy resolution, which a .sh can hardly do),
* processing the shell scripts to replace the version numbers, and
* laying out the contents of the archives.

The script `assemble-cli.sh` is only responsible of invoking the
`cliPack` task in sbt, and create the zip and tgz archives.